refactor: 로또 개수 발급 시 Count로 포장

+++ src/main/java/lotto/view/OutputView.java
@@ -13,8 +13,8 @@ public class OutputView {
     private static final String RESULT_MESSAGE = "%d개 일치 (%d원) - %d개";
 
     public void printBuyStatus(LottoStatusResponseDto lottoStatusResponseDto) {
-        int manualCount = lottoStatusResponseDto.getManualCount();
-        int autoCount = lottoStatusResponseDto.getLottoStatus().size() - manualCount;
+        long manualCount = lottoStatusResponseDto.getManualCount();
+        long autoCount = lottoStatusResponseDto.getLottoStatus().size() - manualCount;
         System.out.printf("수동으로 %d장, 자동으로 %d개를 구매했습니다.", manualCount, autoCount);
         System.out.println();
         for (List<Integer> lottoStatus : lottoStatusResponseDto.getLottoStatus()) {
